==> adjusting start pattern quilt  -

+++ modules/colorutils.py
@@ -95,20 +95,20 @@ def getRandomColor(brtns=1) :
 	a = 255
 	return (r,g,b) 
 
-def getRandomColorHSV(hMin=0, hMax=360, sMin=0, sMax=1, vMin=0, vMax=1) :
+def getRandomColorHSV(hMin=0.0, hMax=360.0, sMin=0.0, sMax=1.0, vMin=0.0, vMax=1.0) :
 
 	# adjust for 360 degrees ranges
 	degreeRange = hMax - hMin
 	
 	if hMin > hMax :
-		degreeRange = 360 - hMin + hMax
-	h = hMin + random.uniform(0, degreeRange)
+		degreeRange = 360.0 - hMin + hMax
+	h = hMin + random.uniform(0.0, degreeRange)
 
-	if (h > 360) : h -= 360
+	if (h > 360.0) : h -= 360.0
 	#h = random.uniform(hMin,hMax)
 	#print(hMin,hMax,degreeRange, h)
-	s = random.uniform(sMin,sMax)
-	v = random.uniform(vMin,vMax)
+	s = random.uniform(sMin, sMax)
+	v = random.uniform(vMin, vMax)
 	return HSVToRGB(h,s,v) 
 
 def randomBaseColor(brtns=1) :
@@ -172,7 +172,7 @@ def HSVToRGB(h,s,v,a=255) :
 	m = v - c
 	rgb  = [r1 + m, g1 + m, b1 + m, a]
 
-	rgbCol = tuple(int(round(i * 255)) for i in rgb)
+	rgbCol = tuple((round(i * 255)) for i in rgb)
 	return rgbCol
 
 def HSLToRGB(h,s,l,a=255) :
